@@ -4,30 +4,33 @@
   "private": true,
   "license": "MIT",
   "homepage": ".",
+  "type": "module",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@testing-library/jest-dom": "^6.6.4",
+    "@testing-library/react": "^16.3.0",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.2.3",
+    "@cypress/vite-dev-server": "^6.0.3",
     "js-yaml": "^4.1.0",
     "path-browserify": "^1.0.1",
     "prism-color-variables": "^1.0.1",
     "prismjs": "^1.29.0",
     "prop-types": "^15.8.1",
-    "react": "^18.2.0",
+    "react": "^19.1.1",
     "react-bootstrap": "^2.7.2",
     "react-diff-view": "^3.2.1",
-    "react-dom": "^18.2.0",
+    "react-dom": "^19.1.1",
     "react-error-boundary": "^4.0.4",
-    "react-scripts": "5.0.1",
     "refractor": "^3.6.0",
+    "vite": "^7.0.6",
+    "@vitejs/plugin-react": "^4.7.0",
+    "vite-plugin-eslint": "^1.8.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "GENERATE_SOURCEMAP=false react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
+    "start": "vite",
+    "build": "vite build",
+    "test": "vitest",
     "lint": "eslint --ext .jsx,.js .",
     "lint:fix": "eslint --ext .jsx,.js --fix .",
     "cypress:open": "cypress open",
@@ -46,8 +49,13 @@
     ]
   },
   "devDependencies": {
-    "cypress": "^13.13.3",
+    "cypress": "^14.5.3",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-plugin-cypress": "^2.14.0"
+    "eslint-plugin-cypress": "^3.6.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "4.3",
+    "eslint-plugin-vitest-globals": "^1.5.0",
+    "jsdom": "^26.1.0",
+    "vitest": "^3.2.4"
   }
 }
