@@ -1,11 +1,18 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
-const { defineConfig } = require('cypress');
+import { defineConfig } from 'cypress';
+import { devServer } from '@cypress/vite-dev-server';
+// eslint-disable-next-line import/extensions
+import viteConfig from './vite.config.js';
 
-module.exports = defineConfig({
+export default defineConfig({
   component: {
-    devServer: {
-      framework: 'create-react-app',
-      bundler: 'webpack',
+    devServer(devServerConfig) {
+      return devServer({
+        ...devServerConfig,
+        framework: 'react',
+        bundler: 'vite',
+        viteConfig,
+      });
     },
     viewportHeight: 800,
     viewportWidth: 500,
